Reg: Fix miscellaneous bugs
Fixed a bug related to using the zero register in certain arithmetic
instructions when the second operand is an immedate. Fixed an issue with
global reads by adding a function to get the *address* of a global,
instead of loading it. Finally, added a separate block for parameters so
that they can be included in register allocation.